Got rid of the redundant googlemaps json calls and fixed the functions so that it works for the except case (that was why it was bugging on other trips)
