+++ emission/net/ext_service/geocoder/nominatim.py
@@ -9,13 +9,18 @@ from builtins import object
 import urllib.request, urllib.parse, urllib.error, urllib.request, urllib.error, urllib.parse
 import logging
 import json
+import re
+import requests
 
 from emission.core.wrapper.trip_old import Coordinate
 from pygeocoder import Geocoder as pyGeo  ## We fall back on this if we have to
 
 try:
     googlemaps_key_file = open("conf/net/ext_service/googlemaps.json")
-    GOOGLE_MAPS_KEY = json.load(googlemaps_key_file)["access_token"]
+    googlemaps_json = json.load(googlemaps_key_file)
+    GOOGLE_MAPS_KEY = googlemaps_json["access_token"]
+    BACKUP_GOOGLE_MAPS_KEY = googlemaps_json["backup_access_token"]
+    NEARBY_URL = googlemaps_json["nearby_base_url"]
 except:
     print("google maps key not configured, falling back to nominatim")
 
@@ -94,6 +99,82 @@ class Geocoder(object):
             return _do_google_reverse(lat, lng) # Just in case
 
 ## Failsafe section
+def check_against_business_location(location='0, 0', address = ''):
+    if not re.compile('^(\-?\d+(\.\d+)?),\s*(\-?\d+(\.\d+)?)$').match(location):
+        raise ValueError('Location Invalid')
+    base_url = NEARBY_URL
+    location = 'location=' + location
+    try:
+        key_string = '&key=' + GOOGLE_MAPS_KEY
+        radius = '&radius=10'
+        url = base_url + location + radius + key_string
+        result = requests.get(url).json()
+        cleaned = result['results']
+        for i in cleaned:
+            #If the street address matches the street address of this business, we return a tuple
+            #signifying success and the business name
+            if address == i['vicinity']:
+                return (True, i['name'])
+        else:
+            return (False, '')
+    except:
+        try:
+            key_string = '&key=' + BACKUP_GOOGLE_MAPS_KEY
+            radius = '&radius=10'
+            url = base_url + location + radius + key_string
+            result = requests.get(url).json()
+            cleaned = result['results']
+            for i in cleaned:
+                if address == i['vicinity']:
+                    return (True, i['name'])
+            else:
+                return (False, '')
+        except:
+            raise ValueError("Something went wrong")
+
+def return_address_from_location_google(location='0,0'):
+    """
+    Creates a Google Maps API call that returns the addresss given a lat, lon
+    """
+    if not re.compile('^(\-?\d+(\.\d+)?),\s*(\-?\d+(\.\d+)?)$').match(location):
+        raise ValueError('Location Invalid')
+    base_url = 'https://maps.googleapis.com/maps/api/geocode/json?'
+    latlng = 'latlng=' + location
+    try:
+        #This try block is for our first 150,000 requests. If we exceed this, use Jack's Token.
+        key_string = '&key=' + GOOGLE_MAPS_KEY
+        url = base_url + latlng + key_string #Builds the url
+        result = requests.get(url).json() #Gets google maps json file
+        cleaned = result['results'][0]['address_components']
+
+
+        #Address to check against value of check_against_business_location
+        chk = cleaned[0]['long_name'] + ' ' + cleaned[1]['long_name'] + ', ' + cleaned[3]['long_name']
+        business_tuple = check_against_business_location(location, chk)
+        
+        if business_tuple[0]: #If true, the lat, lon matches a business location and we return business name
+            address_comp = cleaned[0]['long_name'] + ' ' + cleaned[1]['short_name']
+            return business_tuple[1], cleaned[3]['short_name'], address_comp
+        else: #otherwise, we just return the address
+            return cleaned[0]['long_name'] + ' ' + cleaned[1]['short_name'] + ', ' + cleaned[3]['short_name']
+    except:
+        try:
+            #Use Jack's Token in case of some invalid request problem with other API Token
+            key_string = '&key=' + BACKUP_GOOGLE_MAPS_KEY
+            url = base_url + latlng + key_string #Builds the url
+            result = requests.get(url).json() #Gets google maps json file
+            cleaned = result['results'][0]['address_components']
+            #Address to check against value of check_against_business_location
+            chk = cleaned[0]['long_name'] + ' ' + cleaned[1]['long_name'] + ', ' + cleaned[3]['long_name']
+            business_tuple = check_against_business_location(location, chk)
+            if business_tuple[0]: #If true, the lat, lon matches a business location and we return business name
+                address_comp = cleaned[0]['long_name'] + ' ' + cleaned[1]['short_name'] 
+                return business_tuple[1], cleaned[3]['short_name'], address_comp
+            else: #otherwise, we just return the address
+                return cleaned[0]['long_name'] + ' ' + cleaned[1]['short_name'] + ', ' + cleaned[3]['short_name']
+        except:
+            raise ValueError("Something went wrong")
+
 def _do_google_geo(address):
     geo = pyGeo(GOOGLE_MAPS_KEY)
     results = geo.geocode(address)
